+++ packages/pluginkit/src/v4/server/utils/index.ts
@@ -1,2 +1,2 @@
-export * from './package-json';
+export * from './common';
 export * from './template';
