--- packages/pluginkit/src/v4/server/utils/package-json/index.ts
@@ -1,2 +1,0 @@
-export * from './import';
-export * from './validate';
